update submodule and complex.c

+++ srcs/complex.c
@@ -6,13 +6,13 @@
 /*   By: vblanc <vblanc@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/04 19:14:19 by vblanc            #+#    #+#             */
-/*   Updated: 2025/02/04 19:32:02 by vblanc           ###   ########.fr       */
+/*   Updated: 2025/02/04 19:55:24 by vblanc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/complex.h"
 
-void	ft_set_complex(t_complex *c, int real, int img)
+void	ft_set_complex(t_complex *c, float real, float img)
 {
 	c->real = real;
 	c->img = img;
@@ -35,7 +35,7 @@ t_complex	ft_mul_complex(t_complex c1, t_complex c2)
 	c3.img = c1.real * c2.img + c1.img * c2.real;
 	return (c3);
 }
-int	ft_mod_complex(t_complex c)
+float	ft_mod_complex(t_complex c)
 {
 	return (c.real * c.real + c.img * c.img);
 }
